# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Souq Al Arab Germany API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/signup:
    post:
      operationId: authSignup
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupInput"
      responses:
        "200":
          description: Signed up
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "409":
          description: Email already in use
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /auth/login:
    post:
      operationId: authLogin
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
  /auth/me:
    get:
      operationId: authMe
      tags: [auth]
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    patch:
      operationId: authUpdateProfile
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "400":
          description: Invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /auth/verify-email:
    post:
      operationId: authVerifyEmail
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyEmailInput"
      responses:
        "200":
          description: Email verified, user signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "400":
          description: Invalid or expired code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Account not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /auth/forgot-password:
    post:
      operationId: authForgotPassword
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordInput"
      responses:
        "200":
          description: Reset link queued (always returns ok to avoid leaking accounts)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForgotPasswordResponse"
        "400":
          description: Invalid email
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /auth/reset-password:
    post:
      operationId: authResetPassword
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordInput"
      responses:
        "200":
          description: Password reset successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
        "400":
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
  /auth/resend-verification:
    post:
      operationId: authResendVerification
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResendVerificationInput"
      responses:
        "200":
          description: Sent (always returns ok to avoid leaking accounts)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResendVerificationResponse"
...
  /auth/change-password:
    post:
      operationId: authChangePassword
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordInput"
      responses:
        "200":
          description: Changed
          content:
            application/json:
              schema:
                type: object
                required: [ok]
                properties:
                  ok: { type: boolean }
        "400":
          description: Invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
components:
  schemas:
...
    AuthUser:
      type: object
      required: [id, email, name, phone, city, emailVerified]
      properties:
        id: { type: integer }
        email: { type: string }
        name: { type: string }
        phone: { type: string }
        city: { type: string }
        avatarUrl: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
        emailVerified: { type: boolean }
        followerCount: { type: integer }
        followingCount: { type: integer }
        profileViews: { type: integer }
        adCount: { type: integer }
        devVerificationCode:
          type: string
          description: Returned only outside production while no email provider is connected.
...
    ForgotPasswordInput:
      type: object
      required: [email]
      properties:
        email: { type: string, format: email }
    ForgotPasswordResponse:
      type: object
      required: [ok]
      properties:
        ok: { type: boolean }
        devResetUrl: { type: string }
        devResetToken: { type: string }
    ResetPasswordInput:
      type: object
      required: [token, password]
      properties:
        token: { type: string, minLength: 8 }
        password: { type: string, minLength: 6 }
    OkResponse:
      type: object
      required: [ok]
      properties:
        ok: { type: boolean }
    SignupInput:
      type: object
      required: [email, password, name, phone]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 6 }
        name: { type: string, minLength: 2 }
        phone: { type: string, minLength: 5 }
        city: { type: string }
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 6 }
    VerifyEmailInput:
      type: object
      required: [email, code]
      properties:
        email: { type: string, format: email }
        code: { type: string, minLength: 4, maxLength: 10 }
    ResendVerificationInput:
      type: object
      required: [email]
      properties:
        email: { type: string, format: email }
    ResendVerificationResponse:
      type: object
      required: [ok]
      properties:
        ok: { type: boolean }
        devVerificationCode: { type: string }
...
    UpdateProfileInput:
      type: object
      properties:
        name: { type: string, minLength: 2 }
        phone: { type: string, minLength: 5 }
        city: { type: string }
        avatarUrl: { type: string, nullable: true }
    ChangePasswordInput:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword: { type: string, minLength: 1 }
        newPassword: { type: string, minLength: 6 }
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error: { type: string }
...